+++ litert/vendors/qualcomm/compiler/BUILD
@@ -75,9 +75,6 @@ litert_dynamic_lib(
         "//litert/vendors/qualcomm:common",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core:common",
         "//litert/vendors/qualcomm/core:tensor_pool",
         "//litert/vendors/qualcomm/core/schema:soc_table",
@@ -154,9 +151,6 @@ litert_bin(
         "//litert/vendors/qualcomm:common",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core:common",
         "//litert/vendors/qualcomm/core:tensor_pool",
         "//litert/vendors/qualcomm/core/schema:soc_table",
@@ -270,9 +264,6 @@ litert_lib(
         "//litert/vendors/qualcomm:common",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core:common",
         "//litert/vendors/qualcomm/core:op_code",
         "//litert/vendors/qualcomm/core:tensor_pool",
@@ -362,9 +353,6 @@ litert_lib(
         "//litert/vendors/qualcomm:common",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core:common",
         "@com_google_absl//absl/container:flat_hash_set",
         "@com_google_absl//absl/strings",
