+++ litert/vendors/qualcomm/context_binary_info.cc
@@ -24,7 +24,6 @@
 #include "litert/cc/litert_macros.h"
 #include "litert/vendors/qualcomm/core/wrappers/tensor_wrapper.h"
 #include "litert/vendors/qualcomm/qnn_manager.h"
-#include "litert/vendors/qualcomm/qnn_sdk_version.h"
 #include "QnnCommon.h"  // from @qairt
 #include "QnnTypes.h"  // from @qairt
 #include "System/QnnSystemContext.h"  // from @qairt
@@ -179,8 +178,7 @@ Expected<void> ContextBinaryInfo::Init(
 }
 
 Expected<ContextBinaryInfo> ContextBinaryInfo::Create(
-    QnnManager& qnn, const void* exec_bytecode_ptr,
-    size_t exec_bytecode_size) {
+    QnnManager& qnn, const void* exec_bytecode_ptr, size_t exec_bytecode_size) {
   auto system_context_handle = qnn.CreateSystemContextHandle();
   if (!system_context_handle) {
     return Unexpected(system_context_handle.Error());
