+++ litert/vendors/qualcomm/dispatch/BUILD
@@ -91,9 +91,6 @@ litert_dispatch_api(
         "//litert/vendors/qualcomm:context_binary_info",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core/utils:miscs",
         "//litert/c/internal:litert_scheduling_info",
         "//litert/vendors/qualcomm/core/wrappers:quantize_params_wrapper",
@@ -160,9 +157,6 @@ litert_bin(
         "//litert/vendors/qualcomm:context_binary_info",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core:common",
         "//litert/vendors/qualcomm/core/utils:miscs",
         "//litert/vendors/qualcomm/core/wrappers:quantize_params_wrapper",
