+++ litert/vendors/qualcomm/dispatch/litert_dispatch_invocation_context.h
@@ -31,13 +31,9 @@
 #include "litert/vendors/c/litert_dispatch.h"
 #include "litert/vendors/qualcomm/context_binary_info.h"
 #include "litert/vendors/qualcomm/core/wrappers/tensor_wrapper.h"
-#include "litert/vendors/qualcomm/qnn_handles.h"
+#include "litert/vendors/qualcomm/qnn_manager.h"
 #include "QnnCommon.h"  // from @qairt
 #include "QnnTypes.h"  // from @qairt
-
-namespace litert::qnn {
-class QnnManager;
-}  // namespace litert::qnn
 
 class LiteRtDispatchDeviceContextT;
