+++ litert/vendors/qualcomm/qnn_api_loader.cc
@@ -5,10 +5,13 @@
 
 #include <stdlib.h>
 
+#include <charconv>
 #include <cstdint>
 #include <memory>
 #include <optional>
 #include <string>
+#include <string_view>
+#include <system_error>
 #include <utility>
 
 #include "absl/strings/str_cat.h"  // from @com_google_absl
@@ -29,7 +32,6 @@
 #include "litert/vendors/qualcomm/core/backends/ir_backend.h"
 #include "litert/vendors/qualcomm/core/common.h"
 #include "litert/vendors/qualcomm/qnn_saver_utils.h"
-#include "litert/vendors/qualcomm/qnn_sdk_version.h"
 #include "QnnCommon.h"  // from @qairt
 #include "QnnInterface.h"  // from @qairt
 #include "QnnTypes.h"  // from @qairt
@@ -41,6 +43,46 @@ static constexpr int kRequiredNumProviders{1};
 }
 
 namespace litert::qnn {
+
+Expected<SdkVersion> ParseSdkVersion(const char* build_id) {
+  // Generic parse-failure result.
+  const auto parsing_error =
+      Unexpected(kLiteRtStatusErrorRuntimeFailure, "Failed to parse build ID");
+
+  if (!build_id) return parsing_error;
+
+  std::string_view version_str = build_id;
+
+  // Require the 'v' prefix, then strip it.
+  if (version_str.empty() || version_str.front() != 'v') {
+    return parsing_error;
+  }
+  version_str.remove_prefix(1);
+
+  SdkVersion version{};
+  const char* current = version_str.data();
+  const char* const end = version_str.data() + version_str.size();
+
+  auto parse_component = [&current, &end](int& component) {
+    auto [ptr, ec] = std::from_chars(current, end, component);
+    if (ec != std::errc()) {
+      return false;
+    }
+    current = ptr;
+    return true;
+  };
+
+  // Expect "major.minor.patch".
+  if (!parse_component(version.major)) return parsing_error;
+
+  if (current == end || *current++ != '.') return parsing_error;
+  if (!parse_component(version.minor)) return parsing_error;
+
+  if (current == end || *current++ != '.') return parsing_error;
+  if (!parse_component(version.patch)) return parsing_error;
+
+  return version;
+}
 
 namespace {
