+++ litert/vendors/qualcomm/qnn_backend_test/BUILD
@@ -55,9 +55,6 @@ cc_library(
         "//litert/test:common",
         "//litert/vendors/qualcomm:qnn_api_loader",
         "//litert/vendors/qualcomm:qnn_manager",
-        "//litert/vendors/qualcomm:qnn_context_configs",
-        "//litert/vendors/qualcomm:qnn_handles",
-        "//litert/vendors/qualcomm:qnn_sdk_version",
         "//litert/vendors/qualcomm/core:common",
         "//litert/vendors/qualcomm/core:tensor_pool",
         "//litert/vendors/qualcomm/core/utils:miscs",
