+++ litert/vendors/qualcomm/qnn_backend_test/CMakeLists.txt
@@ -24,7 +24,6 @@ target_link_libraries(qnn_test_utils
         qnn_model
         qnn_api_loader
         qnn_manager
-        qnn_context_configs
     PRIVATE
         GTest::gtest
 )
